unikraft/app,cmd/kraft: Clearly separate app, project and loader options
This commit separates and clearly identifies options into individual
files using similar method and file name semantics.
