+++ cmd/kraft/clean/clean.go
@@ -80,10 +80,10 @@ func (opts *Clean) Run(cmd *cobra.Command, args []string) error {
 	// Initialize at least the configuration options for a project
 	projectOpts, err := app.NewProjectOptions(
 		nil,
-		app.WithWorkingDirectory(workdir),
-		app.WithDefaultConfigPath(),
-		app.WithResolvedPaths(true),
-		app.WithDotConfig(false),
+		app.WithProjectWorkdir(workdir),
+		app.WithProjectDefaultConfigPath(),
+		app.WithProjectResolvedPaths(true),
+		app.WithProjectDotConfig(false),
 	)
 	if err != nil {
 		return err
