+++ cmd/kraft/configure/configure.go
@@ -53,10 +53,10 @@ func (opts *Configure) Run(cmd *cobra.Command, args []string) error {
 	// Initialize at least the configuration options for a project
 	projectOpts, err := app.NewProjectOptions(
 		nil,
-		app.WithWorkingDirectory(workdir),
-		app.WithDefaultConfigPath(),
-		app.WithResolvedPaths(true),
-		app.WithDotConfig(false),
+		app.WithProjectWorkdir(workdir),
+		app.WithProjectDefaultConfigPath(),
+		app.WithProjectResolvedPaths(true),
+		app.WithProjectDotConfig(false),
 	)
 	if err != nil {
 		return err
