+++ cmd/kraft/fetch/fetch.go
@@ -57,10 +57,10 @@ func (opts *Fetch) Run(cmd *cobra.Command, args []string) error {
 	// Initialize at least the configuration options for a project
 	projectOpts, err := app.NewProjectOptions(
 		nil,
-		app.WithWorkingDirectory(workdir),
-		app.WithDefaultConfigPath(),
-		app.WithResolvedPaths(true),
-		app.WithDotConfig(false),
+		app.WithProjectWorkdir(workdir),
+		app.WithProjectDefaultConfigPath(),
+		app.WithProjectResolvedPaths(true),
+		app.WithProjectDotConfig(false),
 	)
 	if err != nil {
 		return err
