+++ cmd/kraft/pkg/pkg.go
@@ -116,10 +116,10 @@ func (opts *Pkg) Run(cmd *cobra.Command, args []string) error {
 	ctx := cmd.Context()
 	projectOpts, err := app.NewProjectOptions(
 		nil,
-		app.WithWorkingDirectory(workdir),
-		app.WithDefaultConfigPath(),
-		app.WithResolvedPaths(true),
-		app.WithDotConfig(true),
+		app.WithProjectWorkdir(workdir),
+		app.WithProjectDefaultConfigPath(),
+		app.WithProjectResolvedPaths(true),
+		app.WithProjectDotConfig(true),
 	)
 	if err != nil {
 		return err
