+++ cmd/kraft/run/run.go
@@ -179,10 +179,10 @@ func (opts *Run) Run(cmd *cobra.Command, args []string) error {
 		target := opts.Target
 		projectOpts, err := app.NewProjectOptions(
 			nil,
-			app.WithWorkingDirectory(workdir),
-			app.WithDefaultConfigPath(),
-			app.WithResolvedPaths(true),
-			app.WithDotConfig(false),
+			app.WithProjectWorkdir(workdir),
+			app.WithProjectDefaultConfigPath(),
+			app.WithProjectResolvedPaths(true),
+			app.WithProjectDotConfig(false),
 		)
 		if err != nil {
 			return err
