+++ cmd/kraft/set/set.go
@@ -110,11 +110,11 @@ func (opts *Set) Run(cmd *cobra.Command, args []string) error {
 	// Initialize at least the configuration options for a project
 	projectOpts, err := app.NewProjectOptions(
 		nil,
-		app.WithWorkingDirectory(workdir),
-		app.WithDefaultConfigPath(),
-		app.WithResolvedPaths(true),
-		app.WithDotConfig(true),
-		app.WithConfig(confOpts),
+		app.WithProjectWorkdir(workdir),
+		app.WithProjectDefaultConfigPath(),
+		app.WithProjectResolvedPaths(true),
+		app.WithProjectDotConfig(true),
+		app.WithProjectConfig(confOpts),
 	)
 	if err != nil {
 		return err
