+++ cmd/kraft/unset/unset.go
@@ -104,11 +104,11 @@ func (opts *Unset) Run(cmd *cobra.Command, args []string) error {
 	// Initialize at least the configuration options for a project
 	projectOpts, err := app.NewProjectOptions(
 		nil,
-		app.WithWorkingDirectory(workdir),
-		app.WithDefaultConfigPath(),
-		app.WithResolvedPaths(true),
-		app.WithDotConfig(true),
-		app.WithConfig(confOpts),
+		app.WithProjectWorkdir(workdir),
+		app.WithProjectDefaultConfigPath(),
+		app.WithProjectResolvedPaths(true),
+		app.WithProjectDotConfig(true),
+		app.WithProjectConfig(confOpts),
 	)
 	if err != nil {
 		return err
