+++ unikraft/app/loader.go
@@ -20,7 +20,6 @@ package app
 import (
 	"fmt"
 	"os"
-	"path/filepath"
 	"strings"
 
 	interp "github.com/compose-spec/compose-go/interpolation"
@@ -42,57 +41,6 @@ const (
 	DefaultOutputDir  = "build"
 	DefaultConfigFile = ".config"
 )
-
-// LoaderOptions supported by Load
-type LoaderOptions struct {
-	// Skip schema validation
-	SkipValidation bool
-	// Skip interpolation
-	SkipInterpolation bool
-	// Skip normalization
-	SkipNormalization bool
-	// Resolve paths
-	ResolvePaths bool
-	// Interpolation options
-	Interpolate *interp.Options
-	// Set project projectName
-	projectName string
-	// Indicates when the projectName was imperatively set or guessed from path
-	projectNameImperativelySet bool
-	// Slice of component options to apply to each loaded component
-	componentOptions []component.ComponentOption
-}
-
-func (o *LoaderOptions) SetProjectName(name string, imperativelySet bool) {
-	o.projectName = normalizeProjectName(name)
-	o.projectNameImperativelySet = imperativelySet
-}
-
-func (o LoaderOptions) GetProjectName() (string, bool) {
-	return o.projectName, o.projectNameImperativelySet
-}
-
-// WithSkipValidation sets the LoaderOptions to skip validation when loading
-// sections
-func WithSkipValidation(opts *LoaderOptions) {
-	opts.SkipValidation = true
-}
-
-func withNamePrecedence(absWorkingDir string, popts *ProjectOptions) func(*LoaderOptions) {
-	return func(lopts *LoaderOptions) {
-		if popts.Name != "" {
-			lopts.SetProjectName(popts.Name, true)
-		} else {
-			lopts.SetProjectName(filepath.Base(absWorkingDir), false)
-		}
-	}
-}
-
-func withComponentOptions(copts ...component.ComponentOption) func(*LoaderOptions) {
-	return func(lopts *LoaderOptions) {
-		lopts.componentOptions = copts
-	}
-}
 
 // Load reads a ConfigDetails and returns a fully loaded configuration
 func Load(details config.ConfigDetails, options ...func(*LoaderOptions)) (*ApplicationConfig, error) {
